requirements.txt
Flask==3.1.3
google-api-python-client==2.194.0
google-auth==2.40.3
google-auth-httplib2==0.3.1
openpyxl==3.1.5
pypdf==6.10.0
python-dotenv==1.1.1
requests==2.32.5
gunicorn==23.0.0
